+++ .github/CHANGELOG.md
@@ -90,6 +90,7 @@
 1. [EFB] Read from VFS JSON5 markup for Ground > Payload page - @2hwk (2Cas)
 1. [FADEC] Fadec rewrite/cleanup/commenting using cpp framework - @frankkopp (Frank Kopp)
 1. [EFCS] Fix ground spoiler retraction after increasing TLA slightly above 0 - @lukecologne (luke)
+1. [FWC] Improved LDG LT memo to take into account light position - @BravoMike99 (bruno_pt99)
 
 ## 0.11.0